feat(web): S-14g -- extract onboarding wizard + channel-setup into onboarding.js
Moves ~320 lines from app.js (1610->1299) into web/modules/onboarding.js:
- First-run onboarding wizard (fetchOnboardingStatus, initOnboarding,
  dismissOnboarding, renderOnboarding, all onbXxx helpers + wireOnboarding)
- showSudoModal (managed-settings pre-flight for Slack connect)
- fallbackCopyToClipboard (legacy browser clipboard helper)
- showSlackManifestModal + initChannelSetup (wires chSlackManifestBtn using
  agentApiName() instead of the private currentAgent from agents.js)

Also fixes a pre-existing bug: showSudoModal was called bare inside agents.js
but was never importable there. Now injected via initAgents() DI
(_showSudoModal?.()), consistent with the other DI callbacks.

Contract test added: onboarding-ui-contract.test.ts guards import wiring,
export surface, DI injection, and HTML marker presence. Suite: 2897/2897 green.
